+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,17 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+0.5.105
++++++++
+
+* Add support for AzureMonitorMetrics Addon (managed prometheus metrics in public preview) for AKS
+
 0.5.104
 ++++++
 
-* Add support for AzureMonitorMetrics Addon (managed prometheus metrics in public preview) for AKS
+* Add support to upgrade or update cluster with managed cluster snapshot. Command is
+    * `az aks upgrade --cluster-snapshot-id <snapshot-id>`
+    * `az aks update --cluster-snapshot-id <snapshot-id>`
 
 0.5.103
 +++++++